+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - 修复 Tabs 切换后 Tab-panel 被销毁的问题
 - 修复 TimePicker 错误的隐藏面板
 - 修复 Table Cell 的样式, #204
+- 修复 Menu default-active 属性不能为空的问题, #200
+- 修复 Menu unique-opened 开启后无法展开子菜单的问题, #200
 - 为 Notification 和 Message 的不同 type 添加独立的调用方法
 - 为 Message Box 和 Dialog 添加 lockScroll 属性，用于定义是否在弹框出现时将 body 滚动锁定
 - 新增 Input textarea 类型的 rows, autosize 属性