+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - 为 Message Box 和 Dialog 添加 lockScroll 属性，用于定义是否在弹框出现时将 body 滚动锁定
 - 新增 Input textarea 类型的 rows, autosize 属性
 - 为 Tree 添加 getCheckedNodes 方法和 node-click、check-change 回调
+- 新增 DatePicker 禁用日期功能 #253
 
 ### 1.0.0-rc.5