+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - 新增 Input textarea 类型的 rows, autosize 属性
 - 为 Tree 添加 getCheckedNodes 方法和 node-click、check-change 回调
 - 新增 DatePicker 禁用日期功能 #253
+- 修复 多选可搜索的 Select 下拉选项自动展开的问题
 
 #### 非兼容性更新