+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 更新日志
 
-### 1.0.0（待发布）
+### 1.0.0-rc.6（待发布）
 
 *2016-XX-XX*
 
@@ -14,6 +14,11 @@
 - 新增 Input textarea 类型的 rows, autosize 属性
 - 为 Tree 添加 getCheckedNodes 方法和 node-click、check-change 回调
 - 新增 DatePicker 禁用日期功能 #253
+- 修复 多选可搜索的 Select 下拉选项自动展开的问题
+- 为 Dialog 添加 top 属性
+- 修复 Menu 组件垂直模式下开启 router 属性会立刻跳转的问题 #295
+- Checkbox change 事件现在只能被人为的交互操作所触发
+- 新增 Checkbox checked 属性
 
 #### 非兼容性更新